Completely refactor gitlab_ci_runner::runner
Before this, the complete hash with all runners was injected into
gitlab_ci_runner::runner. Based on the title of the resource it was then
decided which runner should be extracted from the global hash and merge
with defaults ect.

This was completely refactored in a way that a single
gitlab_ci_runner::runner resource now only holds all data which is need
to configure it but no other.

This also changes the default of gitlab_ci_runner::runners(_defaults) to
an empty hash because I don't see why we need to force user to configure
it.

Move global options to toml generation

This drop the 'builds_dir' and cache_dir because they are runner
specific parameter rather than global ones.

The 'metrics_server' parameter gets dropped as well because it was
deprecated almost 2 years ago.

Furthermore, this commit introduce all missing global configuration
options which are 'log_level', 'log_format' and 'check_interval'.

Move runner configuration from exec to concat::fragment

This uses the foundation laid in
9ff7b039583df596c3d15fea2e62003de5890fdc and moves the runner
configuration from execs to concat::fragments. This was done to be able
to change (parts) of the runner configuration after the inital Puppet
run which is not possible with the exec because there is no config
update logic in the gitlab-runner binary.

Sadly this drops the support for autoregistering a runner on a Gitlab
instance as discussed in
https://github.com/voxpupuli/puppet-gitlab_ci_runner/issues/18.
The registration can be done with Bolt task provide in
https://github.com/voxpupuli/puppet-gitlab_ci_runner/pull/73.

move-to-concat - Move TOML creation to its own function

move-to-concat-deferred - Refactor to use a library

move-to-concat-deferred - bolt-tasks - Add register/unregister Puppet functions

move-to-concat-deferred - Allow $config['token'] to be a deferred function

move-to-concat-deferred - Allow auth and reg token to be retrieved by a Deferred function

move-to-concat-deferred - Don't depend on toml-rb, included the dumper.rb from their

move-to-concat-deferred - Save